# openapi.yaml
openapi: 3.1.0
info:
  title: Bakame API
  version: 1.0.0
  description: |
    Bakame - Production-grade AI platform API

    All endpoints except /health require Bearer token authentication.

    Response Format:
    - Success: { data: <payload>, meta: { requestId: string } }
    - Error: { error: { code: string, message: string, requestId: string } }
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /knowledge:
    get:
      tags: [Knowledge]
      summary: List knowledge items
      operationId: listKnowledge
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Knowledge items

    post:
      tags: [Knowledge]
      summary: Create knowledge item
      operationId: createKnowledge
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title:
                  type: string
                content:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Knowledge item created
        '400':
          $ref: '#/components/responses/ValidationError'
  /knowledge/{id}:
    get:
      tags: [Knowledge]
      summary: Get knowledge item
      operationId: getKnowledgeItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KnowledgeId'
      responses:
        '200':
          description: Knowledge item
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Knowledge]
      summary: Update knowledge item
      operationId: updateKnowledgeItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KnowledgeId'
      responses:
        '200':
          description: Knowledge item updated
  /knowledge/{id}/publish:
    post:
      tags: [Knowledge]
      summary: Submit for review
      operationId: publishKnowledge
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/KnowledgeId'
      responses:
        '200':
          description: Submitted for review
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    KnowledgeId:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - UNAUTHORIZED
                - PERMISSION_DENIED
                - NOT_FOUND
                - RATE_LIMITED
                - INTERNAL_ERROR
            message:
              type: string
            requestId:
              type: string
...